Update patch set 2
Patch Set 2:

Hi Keong
Thanks for your comments.
I have moved to the esr EdgeRules as per your suggestion.
The selflink serves a different purpose, it links the generic-vnf to the instance of the VNF on the VNFM (i.e. the object in the controller), whereas the relationship will link the generic-vnf to the VNFM (i.e. to the controller)

Patch-set: 2