Update patch set 2
Patch Set 2:

As requested, documented the processing steps used to generate the plantuml diagrams for AAI information model reverse engineering discussion.
It is now in the form of a shell script and the generated text files are also submitted.
It is not pretty or optimal but it is at least fully automated.
Takes about 2 mins in my local dev env on my laptop.

Patch-set: 2