Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Is that shell script at the root of the project?

Yes, that's right. It's in the root of the project because it reaches into aai-schema-gen and aai-schema sub-directories during the run.

There might be better places to put it, esp. if we try to integrate it into the pom.xml as an extra maven goal to re-generate diagrams (I was looking at a plantuml-maven plugin but it appears to be unmaintained for the last couple years).

Patch-set: 2