Update patch set 1
Patch Set 1:

Chandra, please review my comments below:

> I think it is redundant to have operational-status in all of these nodes.
> e.g. pserver as well as in vserver.

in a pserver you will have one or more vserver so if one of the vservers goes down then you need to be able to track that.

Similarly the vserver might be moved else where if pserver goes down...

 > We already have operational-status at generic-vnf level. Also, can we use in-maint property in these

having operational-status under generic-vnf is fine but don't forget that a generic-vnf can have one or more vf-modules... so need to be able to track it at the individual vf-module level.

> nodes to indicate whether or not this object is in maintenance mode (maintenance mode = true)

maintenance mode is different... operational status is there to capture the state of the network... not the intent. example you might have maintenance mode set to false and the network resource or compute goes down you need SA to update the operational state regardless of maintenance mode it was put in.

> Patch Set 1:
> 
> I think it is redundant to have operational-status in all of these nodes.
> e.g. pserver as well as in vserver.
> 
> We already have operational-status at generic-vnf level. Also, can we use in-maint property in these
> nodes to indicate whether or not this object is in maintenance mode (maintenance mode = true)
> 
> Thanks
> Chandra

Patch-set: 1