Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (2 comments)

Thanks for the clarification. But I still have a question: how to know if a node is (or is not) top-level? "tunnel-xconnect", for example, is a child under "allotted-resource". How can we tell "tunnel-xconnect" is not a top-level?

Patch-set: 4
Reviewer: Gerrit User 3755 <3755@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>