Update patch set 1
Patch Set 1:

Is this change you raised needed for Guilin? We should not be updating the model this late, as code freeze is today, and even if I merged this change it would have to be staged, released, and charts updated for all the microservices that use schema service.

Patch-set: 1
Reviewer: Gerrit User 2979 <2979@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>