Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Looks like most of the new properties added to vpn-binding node are ids/attributes of another type of node. I suggest to define new node types (slice/access-point) and relate the vpn-binding to the new node type. I think we should avoid adding arbitrary properties to a node but rather use node and edge/relationship.

 
The attributes we try to add to vpn-binding are not "arbitrary". They belong to vpn-binding. srcTpn and dstTpn are related to the existing srcLtp and dstLtp attributes. admin-status, protection-type, and switching-type are network configuration properties to describe and quantify a vpn service. Last, slice-id cannot be replaced by an edge rule, because it is not a reference to another node. Rather, it a network configuration attribute. i.e., SDNC needs to send its value to the domain controller. 

Furthermore, we need to make a balance between clarity and complexity. In Guilin release, our consensus was that we favor the reuse of existing nodes to introducing new ones. As you recall, we decided to reuse "network-route" for 3GPP's endpoint model, and "network-policy" for IETF's TSCi policy model. Both of these two nodes were not originally created for these purposes, but we still added attributes to them to make them fit. Since the AAI schema is already complex, I suggest we keep the same principle in Honolulu release, i.e., avoid adding new nodes and new edge rules. (In fact, this change has already been reviewed in the AAI meetings.)

In the long run, however, AAI should not be used to store network configurations. CCVPN should find a new datastore (e.g., CPS) for that purpose. But that may be a separate discussion.  

Please let me know if this makes sense.

Patch-set: 1