Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> What is the intent of the intent-instance?
> I think we can reuse service-instance node with a specific service-type/role/function values.

Hi Chandra, Thanks a lot for the feedback and comments! After further confirming with CCVPN and E2E Slicing team, we know that service-instance node is a common node of AAI, not only used by CCVPN. So, we will follow your advice to reuse the service-instance node. This gerrit will be withdrawn.

Patch-set: 2
CC: Gerrit User 6289 <6289@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>